+++ src/main/java/com/github/lopoha/coboltruffle/CobolLanguage.java
@@ -1,7 +1,22 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle;
 
 import com.github.lopoha.coboltruffle.builtins.CobolBuiltinNode;
+import com.github.lopoha.coboltruffle.heap.CobolHeap;
+import com.github.lopoha.coboltruffle.heap.HeapPointer;
+import com.github.lopoha.coboltruffle.nodes.CobolEvalRootNode;
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
+import com.github.lopoha.coboltruffle.nodes.CobolMoveNode;
+import com.github.lopoha.coboltruffle.nodes.expression.CobolFunctionLiteralNode;
+import com.github.lopoha.coboltruffle.nodes.expression.CobolStringLiteralNode;
+import com.github.lopoha.coboltruffle.nodes.local.CobolLexicalScope;
+import com.github.lopoha.coboltruffle.parser.CobolBaseListenerImpl;
+import com.github.lopoha.coboltruffle.parser.CobolNodeFactory;
+import com.github.lopoha.coboltruffle.parser.NotImplementedException;
+import com.github.lopoha.coboltruffle.parser.antlr.CobolLexer;
+import com.github.lopoha.coboltruffle.parser.antlr.CobolParser;
 import com.github.lopoha.coboltruffle.parser.common.ParserSettings;
+import com.github.lopoha.coboltruffle.parser.preprocessor.ParserPreprocessor;
+import com.github.lopoha.coboltruffle.runtime.CobolContext;
 import com.oracle.truffle.api.CallTarget;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.RootCallTarget;
@@ -21,50 +36,11 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
+import org.antlr.v4.runtime.CharStream;
+import org.antlr.v4.runtime.CharStreams;
+import org.antlr.v4.runtime.CommonTokenStream;
+import org.antlr.v4.runtime.tree.ParseTreeWalker;
 
-/*
-import com.oracle.truffle.sl.builtins.SLBuiltinNode;
-import com.oracle.truffle.sl.builtins.SLDefineFunctionBuiltin;
-import com.oracle.truffle.sl.builtins.SLNanoTimeBuiltin;
-import com.oracle.truffle.sl.builtins.SLPrintlnBuiltin;
-import com.oracle.truffle.sl.builtins.SLReadlnBuiltin;
-import com.oracle.truffle.sl.builtins.SLStackTraceBuiltin;
-import com.oracle.truffle.sl.nodes.SLEvalRootNode;
-import com.oracle.truffle.sl.nodes.SLTypes;
-import com.oracle.truffle.sl.nodes.controlflow.SLBlockNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLBreakNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLContinueNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLDebuggerNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLIfNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLReturnNode;
-import com.oracle.truffle.sl.nodes.controlflow.SLWhileNode;
-import com.oracle.truffle.sl.nodes.expression.SLAddNode;
-import com.oracle.truffle.sl.nodes.expression.SLBigIntegerLiteralNode;
-import com.oracle.truffle.sl.nodes.expression.SLDivNode;
-import com.oracle.truffle.sl.nodes.expression.SLEqualNode;
-import com.oracle.truffle.sl.nodes.expression.SLFunctionLiteralNode;
-import com.oracle.truffle.sl.nodes.expression.SLInvokeNode;
-import com.oracle.truffle.sl.nodes.expression.SLLessOrEqualNode;
-import com.oracle.truffle.sl.nodes.expression.SLLessThanNode;
-import com.oracle.truffle.sl.nodes.expression.SLLogicalAndNode;
-import com.oracle.truffle.sl.nodes.expression.SLLogicalOrNode;
-import com.oracle.truffle.sl.nodes.expression.SLMulNode;
-import com.oracle.truffle.sl.nodes.expression.SLReadPropertyNode;
-import com.oracle.truffle.sl.nodes.expression.SLStringLiteralNode;
-import com.oracle.truffle.sl.nodes.expression.SLSubNode;
-import com.oracle.truffle.sl.nodes.expression.SLWritePropertyNode;
-import com.oracle.truffle.sl.nodes.local.SLLexicalScope;
-import com.oracle.truffle.sl.nodes.local.SLReadLocalVariableNode;
-import com.oracle.truffle.sl.nodes.local.SLWriteLocalVariableNode;
-import com.oracle.truffle.sl.parser.SLNodeFactory;
-import com.oracle.truffle.sl.parser.SimpleLanguageLexer;
-import com.oracle.truffle.sl.parser.SimpleLanguageParser;
-import com.oracle.truffle.sl.runtime.SLBigNumber;
-import com.oracle.truffle.sl.runtime.SLContext;
-import com.oracle.truffle.sl.runtime.SLFunction;
-import com.oracle.truffle.sl.runtime.SLFunctionRegistry;
-import com.oracle.truffle.sl.runtime.SLNull;
-*/
 
 @TruffleLanguage.Registration(id = CobolLanguage.ID,
                               name = "Cobol",
@@ -99,9 +75,9 @@ public final class CobolLanguage extends TruffleLanguage<CobolContext> {
     List<String> copySearchPath = new ArrayList<>();
     copySearchPath.add("./teststuff/copy");
     ParserSettings parserSettings = new ParserSettings(copySearchPath, programSearchPath);
-    String preprocessed = new Temp().demo_getPreprocessedString("test", parserSettings);
+    String preprocessed = ParserPreprocessor.getPreprocessedString("test", parserSettings);
 
-    Map<String, RootCallTarget> functions = new Temp().demo_processPreprocessed(preprocessed, this);
+    Map<String, RootCallTarget> functions = processPreprocessed(preprocessed, this);
     RootCallTarget main = functions.get("main");
 
     if (main == null) {
@@ -111,6 +87,60 @@ public final class CobolLanguage extends TruffleLanguage<CobolContext> {
 
     RootNode evalMain = new CobolEvalRootNode(this, main, functions);
     return Truffle.getRuntime().createCallTarget(evalMain);
+  }
+
+  /**
+   * TODO: replace method.
+   * @param source the preprocessed source code.
+   * @param cobolLanguage a reference to the cobollanguage.
+   * @return a map with all functions.
+   */
+  public Map<String, RootCallTarget> processPreprocessed(String source,
+                                                              CobolLanguage cobolLanguage) {
+    try {
+      CharStream input = CharStreams.fromString(source);
+      CobolLexer lexer = new CobolLexer(input);
+      CommonTokenStream tokens = new CommonTokenStream(lexer);
+      CobolParser parser = new CobolParser(tokens);
+      CobolParser.FileContext fileContext = parser.file();
+      ParseTreeWalker walker = new ParseTreeWalker();
+      CobolBaseListenerImpl listener = new CobolBaseListenerImpl();
+      walker.walk(listener, fileContext);
+
+      CobolHeap workingStorageHeap = new CobolHeap();
+      workingStorageHeap.addToHeap(listener.workingStorageHeap);
+
+
+
+      CobolNodeFactory cobolNodeFactory = new CobolNodeFactory(cobolLanguage);
+      cobolNodeFactory.startSection("main");
+
+      CobolStringLiteralNode stringConstant = new CobolStringLiteralNode("hello World");
+      CobolFunctionLiteralNode println = new CobolFunctionLiteralNode("display");
+      List<CobolExpressionNode> printlnArgs = new ArrayList<>();
+      printlnArgs.add(stringConstant);
+      cobolNodeFactory.addCall(println, printlnArgs);
+
+      HeapPointer programName = workingStorageHeap.getHeapPointer("PROGRAMNAME");
+      List<CobolExpressionNode> println2Args = new ArrayList<>();
+      println2Args.add(programName);
+      cobolNodeFactory.addCall(println, println2Args);
+
+      CobolMoveNode moveNode = new CobolMoveNode("ABC", programName);
+      cobolNodeFactory.addMove(moveNode);
+      cobolNodeFactory.addCall(println, println2Args);
+
+      HeapPointer copyString = workingStorageHeap.getHeapPointer("COPY-STRING");
+      CobolMoveNode moveNode2 = new CobolMoveNode(copyString, programName);
+      cobolNodeFactory.addMove(moveNode2);
+      cobolNodeFactory.addCall(println, println2Args);
+
+      cobolNodeFactory.finishSection();
+      return cobolNodeFactory.getAllSections();
+
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
   }
 
   /*
