+++ src/main/java/com/github/lopoha/coboltruffle/builtins/CobolBuiltinNode.java
@@ -1,7 +1,7 @@
 package com.github.lopoha.coboltruffle.builtins;
 
-import com.github.lopoha.coboltruffle.parser.CobolException;
-import com.github.lopoha.coboltruffle.parser.CobolExpressionNode;
+import com.github.lopoha.coboltruffle.CobolException;
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
 import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.UnsupportedSpecializationException;
