+++ src/main/java/com/github/lopoha/coboltruffle/heap/CobolVariableSizeMismatch.java
@@ -1,4 +1,4 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle.heap;
 
 // todo: position
 // todo: base parser exception
