+++ src/main/java/com/github/lopoha/coboltruffle/nodes/CobolExpressionNode.java
@@ -1,7 +1,8 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle.nodes;
 
 // modeled / copied from SimpleLanguage
 
+import com.github.lopoha.coboltruffle.parser.NotImplementedException;
 import com.oracle.truffle.api.dsl.TypeSystemReference;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.instrumentation.GenerateWrapper;
