+++ src/main/java/com/github/lopoha/coboltruffle/nodes/CobolRootNode.java
@@ -1,7 +1,8 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle.nodes;
 
 // Modeled after the SLRootNode from SimpleLanguage
 
+import com.github.lopoha.coboltruffle.CobolLanguage;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.frame.FrameDescriptor;
 import com.oracle.truffle.api.frame.VirtualFrame;
