+++ src/main/java/com/github/lopoha/coboltruffle/nodes/CobolStatementNode.java
@@ -1,4 +1,4 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle.nodes;
 
 // modeled / copied from SimpleLanguage
