+++ src/main/java/com/github/lopoha/coboltruffle/nodes/CobolUndefinedSectionRootNode.java
@@ -1,7 +1,6 @@
 package com.github.lopoha.coboltruffle.nodes;
 
-import com.github.lopoha.coboltruffle.parser.CobolLanguage;
-import com.github.lopoha.coboltruffle.parser.CobolRootNode;
+import com.github.lopoha.coboltruffle.CobolLanguage;
 import com.github.lopoha.coboltruffle.runtime.CobolUndefinedNameException;
 import com.oracle.truffle.api.frame.VirtualFrame;
