+++ src/main/java/com/github/lopoha/coboltruffle/nodes/controlflow/CobolSectionBodyNode.java
@@ -1,5 +1,7 @@
-package com.github.lopoha.coboltruffle.parser;
+package com.github.lopoha.coboltruffle.nodes.controlflow;
 
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
+import com.github.lopoha.coboltruffle.nodes.CobolStatementNode;
 import com.github.lopoha.coboltruffle.runtime.CobolNull;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.NodeInfo;
