+++ src/main/java/com/github/lopoha/coboltruffle/nodes/expression/CobolFunctionLiteralNode.java
@@ -1,7 +1,7 @@
-package com.github.lopoha.coboltruffle.expression;
+package com.github.lopoha.coboltruffle.nodes.expression;
 
-import com.github.lopoha.coboltruffle.parser.CobolExpressionNode;
-import com.github.lopoha.coboltruffle.parser.CobolLanguage;
+import com.github.lopoha.coboltruffle.CobolLanguage;
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
 import com.github.lopoha.coboltruffle.runtime.CobolSection;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
