+++ src/main/java/com/github/lopoha/coboltruffle/nodes/local/CobolReadArgumentNode.java
@@ -1,6 +1,6 @@
 package com.github.lopoha.coboltruffle.nodes.local;
 
-import com.github.lopoha.coboltruffle.parser.CobolExpressionNode;
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
 import com.github.lopoha.coboltruffle.runtime.CobolNull;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.profiles.BranchProfile;
