+++ src/main/java/com/github/lopoha/coboltruffle/parser/CobolNodeFactory.java
@@ -1,7 +1,13 @@
 package com.github.lopoha.coboltruffle.parser;
 
-import com.github.lopoha.coboltruffle.expression.CobolInvokeNode;
+import com.github.lopoha.coboltruffle.CobolLanguage;
+import com.github.lopoha.coboltruffle.nodes.CobolExpressionNode;
 import com.github.lopoha.coboltruffle.nodes.CobolMoveNode;
+import com.github.lopoha.coboltruffle.nodes.CobolRootNode;
+import com.github.lopoha.coboltruffle.nodes.CobolStatementNode;
+import com.github.lopoha.coboltruffle.nodes.controlflow.CobolBlockNode;
+import com.github.lopoha.coboltruffle.nodes.controlflow.CobolSectionBodyNode;
+import com.github.lopoha.coboltruffle.nodes.expression.CobolInvokeNode;
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.FrameDescriptor;
@@ -41,11 +47,20 @@ public class CobolNodeFactory {
     allSections.put(this.functionName, Truffle.getRuntime().createCallTarget(rootNode));
   }
 
-  void addMove(CobolMoveNode move) {
+  /**
+   * Todo.
+   * @param move the move to add.
+   */
+  public void addMove(CobolMoveNode move) {
     this.sectionNodes.add(move);
   }
 
-  void addCall(CobolExpressionNode functionNode,
+  /**
+   * todo.
+   * @param functionNode the function that should be called.
+   * @param parameterNodes the parameters for the function.
+   */
+  public void addCall(CobolExpressionNode functionNode,
                                  List<CobolExpressionNode> parameterNodes) {
     if (functionNode == null || parameterNodes.contains(null)) {
       // todo: should this fail?
