+++ src/main/java/com/github/lopoha/coboltruffle/runtime/CobolSectionRegistry.java
@@ -1,6 +1,6 @@
 package com.github.lopoha.coboltruffle.runtime;
 
-import com.github.lopoha.coboltruffle.parser.CobolLanguage;
+import com.github.lopoha.coboltruffle.CobolLanguage;
 import com.github.lopoha.coboltruffle.parser.NotImplementedException;
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.interop.TruffleObject;
